% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModel.R
\name{getHumanReadableModelIdFromModel}
\alias{getHumanReadableModelIdFromModel}
\title{getHumanReadableModelIdFromModel: return human readable verion of model from query}
\usage{
getHumanReadableModelIdFromModel(model)
}
\arguments{
\item{model}{INLA model object that will generatie id from}
}
\value{
Unique String representing model in human readable format
}
\description{
getHumanReadableModelIdFromModel: return human readable verion of model from query
}
