% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModel.R
\name{getHumanReadableModelIdFromQuery}
\alias{getHumanReadableModelIdFromQuery}
\title{getHumanReadableModelIdFromQuery: return human readable verion of model from query}
\usage{
getHumanReadableModelIdFromQuery(query)
}
\arguments{
\item{query}{query object container the observed and the model_type attributes}
}
\value{
Unique String representing model in human readable format
}
\description{
getHumanReadableModelIdFromQuery: return human readable verion of model from query
}
