% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModel.R
\name{getModelIdFromModel}
\alias{getModelIdFromModel}
\title{getModelIdFromModel: function to get model id from a model objejct}
\usage{
getModelIdFromModel(model)
}
\arguments{
\item{model}{INLA object}
}
\description{
getModelIdFromModel: function to get model id from a model objejct
}
