% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModel.R
\name{getModelIdFromQuery}
\alias{getModelIdFromQuery}
\title{getModelIdFromQuery: function to save models and register them in modelDB.csv}
\usage{
getModelIdFromQuery(query)
}
\arguments{
\item{query}{query object container the observed and the model_type attributes}
}
\description{
getModelIdFromQuery: function to save models and register them in modelDB.csv
}
