% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModel.R
\name{getModelQueryObjectFromModel}
\alias{getModelQueryObjectFromModel}
\title{getModelQueryObjectFromModel: return query object from a model.
This is the object we use to generate our unique ids.}
\usage{
getModelQueryObjectFromModel(model, latent = FALSE)
}
\arguments{
\item{model}{= Model object to get query object for}

\item{latent}{= Bool determing if we are saving a latent model or a smooth model}

\item{model_type}{= Model Type string: 'inla_observed' (default) or 'inla_latent'}
}
\value{
An object containing the observed and the model_type fields
}
\description{
getModelQueryObjectFromModel: return query object from a model.
This is the object we use to generate our unique ids.
}
