% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModel.R
\name{getModelQueryObjectFromQuery}
\alias{getModelQueryObjectFromQuery}
\title{getModelQueryObjectFromQuery: Reformate a query object to ensure it is in proper order
before generating id.}
\usage{
getModelQueryObjectFromQuery(query)
}
\arguments{
\item{query}{query object container the observed and the model_type attributes}
}
\value{
An object containing the observed and the model_type fields
}
\description{
getModelQueryObjectFromQuery: Reformate a query object to ensure it is in proper order
before generating id.
}
