% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnModel.R
\name{loadModelFileById}
\alias{loadModelFileById}
\title{loadModelFileById function for getting modeled data}
\usage{
loadModelFileById(filename, model_store_dir = Sys.getenv("MODEL_STORE",
  "/home/rstudio/seattle_flu"), type = "csv")
}
\arguments{
\item{filename}{= At moment we expect full filename in format ID.extension.
This is so in future we can more easily support different model save formats}

\item{model_store_dir}{= directory where models are stored}
}
\value{
model
}
\description{
This function will load a model from the model_store_dir by Id
}
\details{
Future enhancments will add versioning
}
\examples{

}
