% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryModel.R
\name{queryLoadedModel}
\alias{queryLoadedModel}
\title{queryLoadedModel Function querys a model that is already loaded. Used by API}
\usage{
queryLoadedModel(model, outputFile, format = "csv", queryFile = NULL,
  outputDir = Sys.getenv("WORKER_DIR", "/tmp"))
}
\arguments{
\item{model}{modelId to query}

\item{outputFile}{File to save the query result to}

\item{format}{= "csv" (default: pre-defined data-only output), "json" (pre-defined data-only output), or "model" (incidenceMapR model object)}

\item{queryFile}{Path to JSON query to filter the model data for.}

\item{outputDir}{= directory where to save result to}
}
\value{
TRUE if output was written to outputFile
}
\description{
queryLoadedModel Function querys a model that is already loaded. Used by API
}
\examples{

}
