% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryModel.R
\name{queryModelById}
\alias{queryModelById}
\title{queryModelById function for getting modeled data by a known id}
\usage{
queryModelById(modelId, outputFile, queryFile = NULL, format = "csv",
  outputDir = Sys.getenv("WORKER_DIR", "/tmp"))
}
\arguments{
\item{modelId}{modelId to load}

\item{outputFile}{File to save the query result to}

\item{queryFile}{Path to JSON query to filter the model data for.}

\item{format}{= "csv" (default: pre-defined data-only output), "json" (pre-defined data-only output), or "model" (incidenceMapR model object)}

\item{outputDir}{= directory where to save result to}
}
\value{
TRUE if output was written to outputFile
}
\description{
This function loads a cached model object
}
\details{
This function is intended for use by the API server.
}
\examples{

}
