% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnModel.R
\name{returnModel}
\alias{returnModel}
\title{returnModel function for getting modeled data}
\usage{
returnModel(queryIn = jsonlite::toJSON(list(SELECT = list(COLUMN =
  c("samplingLocation", "GEOID")), WHERE = list(COLUMN =
  "samplingLocation", IN = c("kiosk")), GROUP_BY = list(COLUMN =
  c("samplingLocation", "GEOID")), SUMMARIZE = list(COLUMN =
  "samplingLocation", IN = c("kiosk")))), type = "smooth",
  format = "csv", version = "mostRecent",
  cloudDir = "C:/Users/mfamulare/Dropbox (IDM)/SeattleFlu-incidenceMapR/models")
}
\arguments{
\item{queryIn}{JSON query to dbViewR that defines model}

\item{type}{= "smooth" (default), "latent", "effects"}

\item{format}{= "csv" (default: pre-defined data-only output) or "model" (incidenceMapR model object)}

\item{version}{= "mostRecent" (default) or date and time created}

\item{cloudDir}{= directory where models are stored}
}
\value{
dataJSON json with data
}
\description{
This function loads a cached model object containing two datasets
  (1) observed incidence of flu in ILI
  (2) modeled incidence after statistical smoothing
}
\examples{

}
