% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnModelData.R
\name{returnModelData}
\alias{returnModelData}
\title{returnModelData function for getting modeled data}
\usage{
returnModelData(queryIn = jsonlite::toJSON(list(SELECT = list(COLUMN =
  c("samplingLocation", "GEOID")), WHERE = list(COLUMN =
  "samplingLocation", IN = c("kiosk")), GROUP_BY = list(COLUMN =
  c("samplingLocation", "GEOID")), SUMMARIZE = list(COLUMN =
  "samplingLocation", IN = c("kiosk")))), type = "csv",
  cloudDir = "C:/Users/mfamulare/Dropbox (IDM)/SeattleFlu-incidenceMapR/models/")
}
\arguments{
\item{queryIn}{JSON string with description of output data requested from model}
}
\value{
dataJSON json with data
}
\description{
This function loads a cached model object containing two datasets
  (1) observed incidence of flu in ILI
  (2) modeled incidence after statistical smoothing
}
\examples{

}
