% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModel.R
\name{saveModel}
\alias{saveModel}
\title{saveModel: function to save models and register them in modelDB.csv}
\usage{
saveModel(model, modelStoreDir = Sys.getenv("MODEL_STORE",
  "/home/rstudio/seattle_flu/test_model_store"), storeRDS = TRUE)
}
\arguments{
\item{model}{INLA object}
}
\description{
saveModel: function to save models and register them in modelDB.csv
}
