% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModel.R
\name{getModelIdFromModel}
\alias{getModelIdFromModel}
\title{getModelIdFromModel: function to get model id from a model objejct}
\usage{
getModelIdFromModel(mode, model_type = "inla", latent = FALSE)
}
\arguments{
\item{model_type}{String of model type}

\item{latent}{bool for if is latent model}

\item{model}{INLA object}
}
\description{
getModelIdFromModel: function to get model id from a model objejct
}
