% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModel.R
\name{getModelQueryObjectFromModel}
\alias{getModelQueryObjectFromModel}
\title{getModelQueryObjectFromModel: return query object from a model. 
This is the object we use to generate our unique ids.}
\usage{
getModelQueryObjectFromModel(model, model_type = "inla",
  latent = FALSE)
}
\arguments{
\item{model}{= Model object to get query object for}

\item{model_type}{= Model Type string. Default to inla}

\item{latent}{= Bool determing if we are saving a latent model or a smooth model}
}
\value{
An object containing the observed and the model_type fields
}
\description{
getModelQueryObjectFromModel: return query object from a model. 
This is the object we use to generate our unique ids.
}
