% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnModel.R
\name{returnModel}
\alias{returnModel}
\title{returnModel function for getting modeled data}
\usage{
returnModel(queryIn = jsonlite::toJSON(list(SELECT = list(COLUMN =
  c("pathogen", "site_type", "residence_census_tract")), WHERE =
  list(COLUMN = "pathogen", IN = "all"), GROUP_BY = list(COLUMN =
  c("site_type", "residence_census_tract")), SUMMARIZE = list(COLUMN =
  "pathogen", IN = "all"))), type = "observed", version = "latest",
  cloudDir = Sys.getenv("MODEL_STORE",
  "/home/rstudio/seattle_flu/test_model_store"))
}
\arguments{
\item{queryIn}{JSON query to dbViewR that defines model}

\item{type}{= "inla_observed" (default), "inla_latent", "inla"}

\item{version}{= "latest" (default) or ISO date and time created (UTC, like "2019-04-19 22:49:19Z")}

\item{cloudDir}{= directory where models are stored}

\item{format}{= "csv" (default: pre-defined data-only output), "json" (pre-defined data-only output), or "model" (incidenceMapR model object)}
}
\value{
model in requested format
}
\description{
This function loads a cached model object containing two datasets
  (1) observed incidence of flu in ILI
  (2) modeled incidence after statistical smoothing
}
\examples{

}
