% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModel.R
\name{saveModel}
\alias{saveModel}
\title{saveModel: function to save models and register them in modelDB.csv}
\usage{
saveModel(model, cloudDir = "/home/rstudio/seattle_flu/data")
}
\arguments{
\item{model}{INLA object}

\item{db}{dbViewR object}
}
\description{
saveModel: function to save models and register them in modelDB.csv
}
